Show Docker logs if integration test fails (#1587)
* dump log if integration test fails


* update the condition to check exit code

+++ scripts/travis/build-all-in-one-image.sh
@@ -13,6 +13,14 @@ export REPO=jaegertracing/all-in-one
 docker build -f cmd/all-in-one/Dockerfile -t $REPO:latest cmd/all-in-one
 export CID=$(docker run -d -p 16686:16686 -p 5778:5778 $REPO:latest)
 make integration-test
+
+if [ $? -ne 0 ] ; then
+    echo "---- integration test failed unexpectedly ----"
+    echo "--- check the docker log below for details ---"
+    docker logs $CID
+    exit 1
+fi
+
 docker kill $CID
 
 # Only push the docker container to Docker Hub for master branch
